+++ api-config.js
@@ -17,8 +17,16 @@ module.exports = {
     needCommonUtil: false,
     pkgInfo: [
       {
-        version: '1.0.3',
+        version: '1.0.5',
         name: '@uni/env',
+        exports: {
+          web: './es/web/index.js',
+          'wechat-miniprogram': './es/wechat-miniprogram/index.js',
+          'bytedance-microapp': './es/bytedance-microapp/index.js',
+          miniapp: './es/ali-miniapp/index.js',
+          require: './lib/index.js',
+          default: './es/index.js',
+        },
       },
     ],
   },
@@ -100,7 +108,7 @@ module.exports = {
     path: 'src/packages/file/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.4',
+        version: '1.0.7',
         name: '@uni/file',
       },
     ],