+++ src/packages/file/src/web/upload.ts
@@ -1,4 +1,4 @@
-import { UploadOptions } from '../types';
+import { UploadHeadersReceivedCallback, UploadOptions, UploadProgressUpdateCallback, UploadProgressUpdateRes, UploadTask } from '../types';
 import { normalize } from '../common';
 import { CONTAINER_NAME } from '@utils/constant';
 
@@ -31,82 +31,128 @@ function base64toFile(dataUrl: string, fileName = '') {
   return file;
 }
 
-function validateStatus(status: number) {
-  return (status >= 200 && status < 300) || status === 304;
+function getHeaderMap(xhr: XMLHttpRequest) {
+  const headers = xhr.getAllResponseHeaders();
+  const arr = headers.trim().split(/[\r\n]+/);
+  const headerMap: { [x: string]: string } = {};
+  arr.forEach((line) => {
+    const parts = line.split(': ');
+    const _header = parts.shift();
+    const value = parts.join(': ');
+    headerMap[_header as string] = value;
+  });
+  return headerMap;
 }
 
-function uploadFile(param: UploadOptions) {
-  // return new Promise((resolve, reject) => {
-  const file = typeof param.filePath === 'string' ? base64toFile(param.filePath) : param.filePath;
+function uploadFile(param: UploadOptions): UploadTask {
+  const { filePath, formData, success, fail, complete } = param;
+  const file = typeof filePath === 'string' ? base64toFile(filePath) : filePath;
   const body = new FormData();
-  if (param.formData) {
-    for (const key in param.formData) {
-      if (Object.prototype.hasOwnProperty.call(param.formData, key)) {
-        body.append(key, param.formData[key]);
+  if (formData) {
+    for (const key in formData) {
+      if (Object.prototype.hasOwnProperty.call(formData, key)) {
+        body.append(key, formData[key]);
       }
     }
   }
   body.append(param.fileName || 'file', file);
   const header = {
     Accept: 'application/json, text/plain, */*',
-    ...param.header || {},
+    ...(param.header || {}),
   };
-    // initialize xhr
+  // initialize xhr
   const xhr = new XMLHttpRequest();
+  const headersReceivedCallback: UploadHeadersReceivedCallback[] = [];
   xhr.onreadystatechange = () => {
-    if (!xhr || xhr.readyState !== 4) {
-      return;
+    // header received
+    if (xhr && xhr.readyState === 2 && headersReceivedCallback.length > 0) {
+      const headers = getHeaderMap(xhr);
+      headersReceivedCallback.forEach((x) => x(headers));
     }
-    // fail
-    if (!validateStatus(xhr.status)) {
-      param.fail && param.fail({ errMsg: '上传失败' });
-      param.complete && param.complete({ errMsg: '上传失败' });
-      return;
-      // return reject(xhr.status);
-    }
-    // process response headers
-    const headers = xhr.getAllResponseHeaders();
-    const arr = headers.trim().split(/[\r\n]+/);
-    const headerMap = {};
-    arr.forEach((line) => {
-      const parts = line.split(': ');
-      const _header = parts.shift();
-      const value = parts.join(': ');
-      headerMap[_header as string] = value;
-    });
-    let data: any = xhr.response;
-    try {
-      data = JSON.parse(data);
-    } catch (e) {
-      // ignore
-    }
-    // success
+  };
+  // Success
+  xhr.onload = () => {
     const result = {
-      data,
+      data: xhr.response,
       statusCode: xhr.status,
-      header: headerMap,
+      errMsg: 'uploadFile:ok',
+      header: getHeaderMap(xhr),
     };
-    param.success && param.success(result);
-    param.complete && param.complete(result);
-    // resolve(result);
+    success && success(result);
+    complete && complete(result);
+  };
+  // Errors
+  xhr.onerror = () => {
+    fail && fail({ errMsg: 'uploadFile:fail error' });
+    complete && complete({ errMsg: 'uploadFile:fail error' });
+  };
+  xhr.onabort = () => {
+    fail && fail({ errMsg: 'uploadFile:fail abort' });
+    complete && complete({ errMsg: 'uploadFile:fail abort' });
+  };
+  // Progress event
+  const progressCallback: UploadProgressUpdateCallback[] = [];
+  xhr.onprogress = (evt) => {
+    if (progressCallback.length === 0) {
+      return;
+    }
+    const res: UploadProgressUpdateRes = {
+      progress: 0,
+      totalBytesSent: 0,
+      totalBytesExpectedToSend: 0,
+    };
+    if (evt.lengthComputable) {
+      res.totalBytesSent = evt.loaded;
+      res.totalBytesExpectedToSend = evt.total;
+      res.progress = Math.floor((100 * evt.loaded) / evt.total);
+    }
+    progressCallback.forEach((x) => x(res));
   };
   // check if need add withCredentials
-  if (param.url.indexOf(window.location.host) === -1) {
-    xhr.withCredentials = true;
+  if (typeof param.withCredentials === 'undefined') {
+    if (param.url.indexOf(window.location.host) === -1) {
+      xhr.withCredentials = true;
+    }
+  } else {
+    xhr.withCredentials = param.withCredentials;
   }
   xhr.open('POST', param.url, true);
-  // Object.keys(header).forEach((key) => {
-  //   xhr.setRequestHeader(key, String(header[key]));
-  // });
   for (const headerKey in header) {
     if (Object.prototype.hasOwnProperty.call(header, headerKey)) {
       xhr.setRequestHeader(headerKey, String(header[headerKey]));
     }
   }
+  // timeout
+  if (param.timeout) {
+    xhr.timeout = param.timeout;
+    xhr.ontimeout = () => {
+      fail && fail({ errMsg: 'uploadFile:fail timeout' });
+      complete && complete({ errMsg: 'uploadFile:fail timeout' });
+    };
+  }
   // send request
   xhr.send(body);
-  return { abort: () => {} };
-  // });
+  return {
+    abort() {
+      xhr.abort();
+    },
+    onProgressUpdate(cb) {
+      progressCallback.push(cb);
+    },
+    offProgressUpdate(cb) {
+      if (progressCallback.indexOf(cb) !== -1) {
+        progressCallback.splice(progressCallback.indexOf(cb), 1);
+      }
+    },
+    onHeadersReceived(cb) {
+      headersReceivedCallback.push(cb);
+    },
+    offHeadersReceived(cb) {
+      if (headersReceivedCallback.indexOf(cb) !== -1) {
+        headersReceivedCallback.splice(headersReceivedCallback.indexOf(cb), 1);
+      }
+    },
+  };
 }
 
 export default normalize.upload(uploadFile, CONTAINER_NAME.WEB);